Avoid U128
