+++ esp-hal/src/interrupt/xtensa.rs
@@ -2,6 +2,7 @@ use xtensa_lx::interrupt::{self, InterruptNumber};
 use xtensa_lx_rt::exception::Context;
 
 pub use self::vectored::*;
+use super::InterruptStatus;
 use crate::{
     peripherals::{self, Interrupt},
     Cpu,
@@ -149,66 +150,86 @@ pub fn clear(_core: Cpu, which: CpuInterrupt) {
 }
 
 /// Get status of peripheral interrupts
-pub fn get_status(core: Cpu) -> u128 {
+#[cfg(large_intr_status)]
+pub fn get_status(core: Cpu) -> InterruptStatus {
     unsafe {
-        let status = match core {
-            Cpu::ProCpu => {
-                ((*core0_interrupt_peripheral())
+        match core {
+            Cpu::ProCpu => InterruptStatus::from(
+                (*core0_interrupt_peripheral())
                     .pro_intr_status_0()
                     .read()
-                    .bits() as u128)
-                    | ((*core0_interrupt_peripheral())
-                        .pro_intr_status_1()
-                        .read()
-                        .bits() as u128)
-                        << 32
-                    | ((*core0_interrupt_peripheral())
-                        .pro_intr_status_2()
-                        .read()
-                        .bits() as u128)
-                        << 64
-            }
+                    .bits(),
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_1()
+                    .read()
+                    .bits(),
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_2()
+                    .read()
+                    .bits(),
+            ),
             #[cfg(multi_core)]
-            Cpu::AppCpu => {
-                ((*core1_interrupt_peripheral())
+            Cpu::AppCpu => InterruptStatus::from(
+                (*core1_interrupt_peripheral())
                     .app_intr_status_0()
                     .read()
-                    .bits() as u128)
-                    | ((*core1_interrupt_peripheral())
-                        .app_intr_status_1()
-                        .read()
-                        .bits() as u128)
-                        << 32
-                    | ((*core1_interrupt_peripheral())
-                        .app_intr_status_2()
-                        .read()
-                        .bits() as u128)
-                        << 64
-            }
-        };
+                    .bits(),
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_1()
+                    .read()
+                    .bits(),
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_2()
+                    .read()
+                    .bits(),
+            ),
+        }
+    }
+}
 
-        #[cfg(esp32s3)]
-        let status = match core {
-            Cpu::ProCpu => {
-                status
-                    | ((*core0_interrupt_peripheral())
-                        .pro_intr_status_3()
-                        .read()
-                        .bits() as u128)
-                        << 96
-            }
+/// Get status of peripheral interrupts
+#[cfg(very_large_intr_status)]
+pub fn get_status(core: Cpu) -> InterruptStatus {
+    unsafe {
+        match core {
+            Cpu::ProCpu => InterruptStatus::from(
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_0()
+                    .read()
+                    .bits(),
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_1()
+                    .read()
+                    .bits(),
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_2()
+                    .read()
+                    .bits(),
+                (*core0_interrupt_peripheral())
+                    .pro_intr_status_3()
+                    .read()
+                    .bits(),
+            ),
             #[cfg(multi_core)]
-            Cpu::AppCpu => {
-                status
-                    | ((*core1_interrupt_peripheral())
-                        .app_intr_status_3()
-                        .read()
-                        .bits() as u128)
-                        << 96
-            }
-        };
-
-        status
+            Cpu::AppCpu => InterruptStatus::from(
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_0()
+                    .read()
+                    .bits(),
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_1()
+                    .read()
+                    .bits(),
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_2()
+                    .read()
+                    .bits(),
+                (*core1_interrupt_peripheral())
+                    .app_intr_status_3()
+                    .read()
+                    .bits(),
+            ),
+        }
     }
 }
 
@@ -310,8 +331,12 @@ mod vectored {
     }
 
     /// Get the interrupts configured for the core
-    #[inline]
-    fn get_configured_interrupts(core: Cpu, mut status: u128) -> [u128; 8] {
+    #[inline(always)]
+    fn get_configured_interrupts(
+        core: Cpu,
+        status: InterruptStatus,
+        level: u32,
+    ) -> InterruptStatus {
         unsafe {
             let intr_map_base = match core {
                 Cpu::ProCpu => (*core0_interrupt_peripheral()).pro_mac_intr_map().as_ptr(),
@@ -319,22 +344,22 @@ mod vectored {
                 Cpu::AppCpu => (*core1_interrupt_peripheral()).app_mac_intr_map().as_ptr(),
             };
 
-            let mut levels = [0u128; 8];
+            let mut res = InterruptStatus::empty();
 
-            while status != 0 {
-                let interrupt_nr = status.trailing_zeros();
+            for interrupt_nr in status.iterator() {
                 let i = interrupt_nr as isize;
                 let cpu_interrupt = intr_map_base.offset(i).read_volatile();
                 // safety: cast is safe because of repr(u32)
                 let cpu_interrupt: CpuInterrupt =
                     core::mem::transmute::<u32, CpuInterrupt>(cpu_interrupt);
-                let level = cpu_interrupt.level() as u8 as usize;
+                let int_level = cpu_interrupt.level() as u8 as u32;
 
-                levels[level] |= 1 << i;
-                status &= !(1u128 << interrupt_nr);
+                if int_level == level {
+                    res.set(interrupt_nr as u16);
+                }
             }
 
-            levels
+            res
         }
     }
 
@@ -437,6 +462,8 @@ mod vectored {
 
     #[ram]
     unsafe fn handle_interrupts(level: u32, save_frame: &mut Context) {
+        let core = crate::get_core();
+
         let cpu_interrupt_mask =
             interrupt::get() & interrupt::get_mask() & CPU_INTERRUPT_LEVELS[level as usize];
 
@@ -458,30 +485,23 @@ mod vectored {
             // for edge interrupts cannot rely on the interrupt status
             // register, therefore call all registered
             // handlers for current level
-            let interrupt_levels =
-                get_configured_interrupts(crate::get_core(), chip_specific::INTERRUPT_EDGE);
-            let interrupt_mask = interrupt_levels[level as usize];
-            let mut interrupt_mask = interrupt_mask & chip_specific::INTERRUPT_EDGE;
-            loop {
-                let interrupt_nr = interrupt_mask.trailing_zeros();
-                if let Ok(interrupt) = peripherals::Interrupt::try_from(interrupt_nr as u16) {
-                    handle_interrupt(level, interrupt, save_frame)
-                } else {
-                    break;
-                }
-                interrupt_mask &= !(1u128 << interrupt_nr);
+            let configured_interrupts =
+                get_configured_interrupts(core, chip_specific::INTERRUPT_EDGE, level);
+
+            for interrupt_nr in configured_interrupts.iterator() {
+                // Don't use `Interrupt::try_from`. It's slower and placed in flash
+                let interrupt = unsafe { core::mem::transmute(interrupt_nr as u16) };
+                handle_interrupt(level, interrupt, save_frame);
             }
         } else {
-            // finally check periperal sources and fire of handlers from pac
+            // finally check peripheral sources and fire of handlers from pac
             // peripheral mapped interrupts are cleared by the peripheral
-            let status = get_status(crate::get_core());
-            let interrupt_levels = get_configured_interrupts(crate::get_core(), status);
-            let interrupt_mask = status & interrupt_levels[level as usize];
-            let interrupt_nr = interrupt_mask.trailing_zeros();
+            let status = get_status(core);
+            let configured_interrupts = get_configured_interrupts(core, status, level);
 
-            // Interrupt::try_from can fail if interrupt already de-asserted:
-            // silently ignore
-            if let Ok(interrupt) = peripherals::Interrupt::try_from(interrupt_nr as u16) {
+            for interrupt_nr in configured_interrupts.iterator() {
+                // Don't use `Interrupt::try_from`. It's slower and placed in flash
+                let interrupt = unsafe { core::mem::transmute(interrupt_nr as u16) };
                 handle_interrupt(level, interrupt, save_frame);
             }
         }
@@ -511,8 +531,11 @@ mod vectored {
     #[cfg(esp32)]
     mod chip_specific {
         use super::*;
-        pub const INTERRUPT_EDGE: u128 =
-    0b_0000_0000_0000_0000_0000_0000_0000_0000__0000_0000_0000_0000_0000_0000_0000_0011_1111_1100_0000_0000_0000_0000_0000_0000__0000_0000_0000_0000_0000_0000_0000_0000;
+        pub const INTERRUPT_EDGE: InterruptStatus = InterruptStatus::from(
+            0b0000_0000_0000_0000_0000_0000_0000_0011,
+            0b1111_1100_0000_0000_0000_0000_0000_0000,
+            0b0000_0000_0000_0000_0000_0000_0000_0000,
+        );
         #[inline]
         pub fn interrupt_is_edge(interrupt: Interrupt) -> bool {
             use peripherals::Interrupt::*;
@@ -534,8 +557,11 @@ mod vectored {
     #[cfg(esp32s2)]
     mod chip_specific {
         use super::*;
-        pub const INTERRUPT_EDGE: u128 =
-    0b_0000_0000_0000_0000_0000_0000_0000_0000__0000_0000_0000_0000_0000_0011_1011_1111_1100_0000_0000_0000_0000_0000_0000_0000__0000_0000_0000_0000_0000_0000_0000_0000;
+        pub const INTERRUPT_EDGE: InterruptStatus = InterruptStatus::from(
+            0b0000_0000_0000_0000_0000_0011_1011_1111,
+            0b1100_0000_0000_0000_0000_0000_0000_0000,
+            0b0000_0000_0000_0000_0000_0000_0000_0000,
+        );
         #[inline]
         pub fn interrupt_is_edge(interrupt: Interrupt) -> bool {
             use peripherals::Interrupt::*;
@@ -559,7 +585,7 @@ mod vectored {
     #[cfg(esp32s3)]
     mod chip_specific {
         use super::*;
-        pub const INTERRUPT_EDGE: u128 = 0;
+        pub const INTERRUPT_EDGE: InterruptStatus = InterruptStatus::empty();
         #[inline]
         pub fn interrupt_is_edge(_interrupt: Interrupt) -> bool {
             false
