+++ esp-metadata/devices/esp32.toml
@@ -60,6 +60,7 @@ symbols = [
     "wifi",
     "psram",
     "timg_timer1",
+    "large_intr_status",
 
     # ROM capabilities
     "rom_crc_le",
