+++ esp-metadata/devices/esp32s2.toml
@@ -58,6 +58,7 @@ symbols = [
     "psram",
     "ulp_riscv_core",
     "timg_timer1",
+    "large_intr_status",
 
     # ROM capabilities
     "rom_crc_le",
